+++ bitcoin-de-sdk-value/src/main/kotlin/de/nykon/bitcoin/sdk/value/showAccountInfo/Eth.kt
@@ -1,7 +1,9 @@
 package de.nykon.bitcoin.sdk.value.showAccountInfo
 
+import java.math.BigDecimal
+
 data class Eth(
-    val available_amount: Double,
-    val reserved_amount: Double,
-    val total_amount: Double
+        val available_amount: BigDecimal,
+        val reserved_amount: BigDecimal,
+        val total_amount: BigDecimal
 )
